books/requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.1
MarkupSafe==2.1.3
numpy==1.26.3
packaging==23.2
pandas==2.2.0
psycopg2==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1
